+++ fsl/tests/test_wrappers/test_wrappers.py
@@ -653,3 +653,10 @@ def test_get_standard():
     with testenv('fsl_get_standard') as exe:
         assert fw.get_standard('brain', 'T1')     == f'{exe} brain T1'
         assert fw.get_standard('whole_head', r=2) == f'{exe} whole_head -r 2'
+
+
+def test_makerot():
+    with testenv('makerot') as exe:
+        assert fw.makerot(90).stdout[0] == f'{exe} -t 90'
+        assert fw.makerot(90, cov='cov', centre=[10, 10, 10]).stdout[0] == \
+                          f'{exe} -t 90 --cov=cov --centre=10,10,10'